完成学生输入厘米自动转换成米

+++ database/migrations/2015_09_16_133002_CreateStudentsTable.php
@@ -16,7 +16,7 @@ class CreateStudentsTable extends Migration
             $table->increments('id');
             $table->unsignedTinyInteger('num');
             $table->string('sex');
-            $table->unsignedTinyInteger('height');
+            $table->float('height');
             $table->unsignedTinyInteger('weight');
             $table->timestamps();
         });
